Rename CSI volumes package
Renames the github.com/docker/swarmkit/manager/volumes package to
github.com/docker/swarmkit/manager/csi, which more accurately reflects
what the purpose of that package is.
