+++ manager/csi/convert.go
@@ -1,4 +1,4 @@
-package volumes
+package csi
 
 import (
 	"github.com/container-storage-interface/spec/lib/go/csi"
