+++ manager/csi/csi_mock_test.go
@@ -1,4 +1,4 @@
-package volumes
+package csi
 
 import (
 	"context"
